+++ dubbo-samples-triple-reactor/src/main/java/org/apache/dubbo/samples/triple/reactor/util/EmbeddedZooKeeper.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.dubbo.samples.reactor.util;
+package org.apache.dubbo.samples.triple.reactor.util;
 
 import org.apache.zookeeper.server.ServerConfig;
 import org.apache.zookeeper.server.ZooKeeperServerMain;
