+++ dubbo-samples-triple-reactor/src/test/java/org/apache/dubbo/samples/triple/reactor/ReactorServerTest.java
@@ -1,13 +1,27 @@
-package org.apache.dubbo.samples.reactor;
+/*
+ *  Licensed to the Apache Software Foundation (ASF) under one or more
+ *  contributor license agreements.  See the NOTICE file distributed with
+ *  this work for additional information regarding copyright ownership.
+ *  The ASF licenses this file to You under the Apache License, Version 2.0
+ *  (the "License"); you may not use this file except in compliance with
+ *  the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package org.apache.dubbo.samples.triple.reactor;
 
 import org.apache.dubbo.common.constants.CommonConstants;
 import org.apache.dubbo.config.ApplicationConfig;
 import org.apache.dubbo.config.ReferenceConfig;
 import org.apache.dubbo.config.RegistryConfig;
 import org.apache.dubbo.config.bootstrap.DubboBootstrap;
-import org.apache.dubbo.sample.reactor.GreeterReply;
-import org.apache.dubbo.sample.reactor.GreeterRequest;
-import org.apache.dubbo.sample.reactor.GreeterService;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import reactor.core.publisher.Flux;
